pyproject.toml
[project]
name = "chat-ia-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Vinicius Faria",email = "vinicius.f4ria@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.61.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
